requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
pydantic